Refactor compare runner parallel tests into focused modules
